backend/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain_google_genai
gitpython
docker
pydantic
google-genai
python-dotenv
pytest
pytest-asyncio
python-multipart
